backend/requirements.txt
# Backend Requirements
# Core
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0

# Database
# sqlite3 is built-in with Python

# Utilities
requests==2.31.0
python-dateutil==2.8.2
openpyxl==3.1.5
pandas>=2.2.0

# Development
pytest==7.4.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
